animation

+++ src/cs5004/animator/model/Point.java
@@ -1,18 +1,18 @@
 package cs5004.animator.model;
 
 public class Point {
-  private double x;
-  private double y;
-  public Point(double x, double y) {
+  private int x;
+  private int y;
+  public Point(int x, int y) {
     this.x = x;
     this.y = y;
   }
 
-  public double getX() {
+  public int getX() {
     return x;
   }
 
-  public double getY() {
+  public int getY() {
     return y;
   }
 }
